Extract print ticket spec to a separate file
